+++ aws/resources.go
@@ -478,6 +478,11 @@ func albTargetGroupAttachments(ctx context.Context, a *aws, resourceType string,
 		}
 
 		for _, t := range targetHealths {
+			// As this are the required values to get the resource
+			// we validate that they are present
+			if t.Target == nil || i.TargetGroupArn == nil {
+				continue
+			}
 			r, err := initializeResource(a, fmt.Sprintf("%s_%d_%s", *t.Target.Id, *t.Target.Port, *i.TargetGroupArn), resourceType)
 			if err != nil {
 				return nil, err